feat: change wallet connecting logic
